Update PowerShell - DISM, Add-WindowsDriver, Mount-DiskImage, Mount-WindowsImage (Add drivers to Windows bootable .iso images).ps1

+++ powershell/PowerShell - DISM, Add-WindowsDriver, Mount-DiskImage, Mount-WindowsImage (Add drivers to Windows bootable .iso images).ps1
@@ -74,13 +74,8 @@ If ((Test-Path ("${Install_Wim}")) -Eq $False) {
 						$KeepImage = $True;
 						${WimInfoIndex} = $EachIndex;
 						Write-Host "`${WimInfoIndex} = ${WimInfoIndex}";
-					} Else {
 					}
 				}
-				Write-Host "------------------------------------------------------------";
-				Write-Host "`${EachIndex} = ${EachIndex}";
-				Write-Host "Calling  [ 		Get-WindowsImage -ImagePath ("${Install_Esd}") -Index (${EachIndex}); ] ..."
-				Get-WindowsImage -ImagePath ("${Install_Esd}") -Index (${EachIndex});
 				If (${KeepImage} -Eq $False) {
 					$InvalidWimIndices += $EachIndex;
 				}
@@ -93,23 +88,29 @@ If ((Test-Path ("${Install_Wim}")) -Eq $False) {
 			${WimInfoIndex} = 1;
 		}
 
+		Write-Host "WimInfoIndex = ${WimInfoIndex} ";
+		Write-Host "Calling  [ Get-WindowsImage -ImagePath (`"${Install_Wim}`") -Index (${WimInfoIndex}); ] ...";
+		Get-WindowsImage -ImagePath ("${Install_Wim}") -Index (${WimInfoIndex});
+
 		<# Export the image by creating/updating the "Install.wim" windows image-file #>
 		<#   > Note: this process often requires a few (~2-3) minutes to complete, and may take longer if you've added many more drivers to the customized Windows image #>
+		Write-Host "Calling  [ DISM /Export-Image /SourceImageFile:`"${Install_Esd}`" /SourceIndex:${WimInfoIndex} /DestinationImageFile:`"${Install_Wim}`" /Compress:max /CheckIntegrity; ] ...";
 		DISM /Export-Image /SourceImageFile:"${Install_Esd}" /SourceIndex:${WimInfoIndex} /DestinationImageFile:"${Install_Wim}" /Compress:max /CheckIntegrity;
-
-		<# Remove various Windows images from the image to-be-exported (education version, home version, etc.) #>
-		<#   > Note: This is performed separately because (at the time of writing this) I believe the Remove-WindowsImage must refer to the "install.wim" and not the "install.esd" file #>
-		$InvalidWimIndices | ForEach-Object {
-			$EachImageIndex = $_;
-			Remove-WindowsImage -ImagePath ("${Install_Wim}") -Index (${EachImageIndex}) -CheckIntegrity;
-		}
-
 
 		<# Double-check to ensure that this image is the one you want #>
 		Get-WindowsImage -ImagePath ("${Install_Wim}") -Index (${WimInfoIndex});
+
 	}
 }
 
+<# Remove various Windows images from the image to-be-exported (education version, home version, etc.) #>
+<#   > Note: This is performed separately because (at the time of writing this) I believe the Remove-WindowsImage must refer to the "install.wim" and not the "install.esd" file #>
+If ($False) {
+$InvalidWimIndices | ForEach-Object {
+	$EachImageIndex = $_;
+	Remove-WindowsImage -ImagePath ("${Install_Wim}") -Index (${EachImageIndex}) -CheckIntegrity;
+}
+}
 
 
 If ($True) {
